chore: PackageSelector -> VersionSelector
